Add run-ci-tests, the generic test runner
This script should be started with zero, one or two arguments.
The first argument is the location of the source directory. If left
out it defaults to the current directory. The second argument is the
ci info directory. If left out it is '$src_directory/ci-info'.

The ci info directory describes all the debian and pip dependencies
required and has a number of ci scripts that should be executed.

The run-ci-tests script creates a scratch directory, installs debian
packages, deploys virtualenv, installs python packages and then
proceeds to running each of the ci scripts

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>